Conditional compilation openssl
